+++ packages/echo/text-model/src/text-model.test.ts
@@ -5,6 +5,7 @@
 import expect from 'expect';
 import { it as test } from 'mocha';
 
+import { PublicKey } from '@dxos/crypto';
 import { MockFeedWriter } from '@dxos/echo-protocol';
 import { ModelFactory, TestRig } from '@dxos/model-factory';
 
@@ -30,13 +31,13 @@ describe('TextModel', () => {
 
   test('snapshot', async () => {
     const modelFactory = new ModelFactory().registerModel(TextModel);
-    const model1 = modelFactory.createModel<TextModel>(TextModel.meta.type, 'test', {}, new MockFeedWriter());
+    const model1 = modelFactory.createModel<TextModel>(TextModel.meta.type, 'test', {}, PublicKey.random(), new MockFeedWriter());
 
     model1.model.insert(0, 'Hello World!');
 
     const snapshot = model1.createSnapshot();
 
-    const model2 = modelFactory.createModel<TextModel>(TextModel.meta.type, 'test', snapshot, new MockFeedWriter());
+    const model2 = modelFactory.createModel<TextModel>(TextModel.meta.type, 'test', snapshot, PublicKey.random(), new MockFeedWriter());
 
     expect(model2.model.textContent).toBe('Hello World!');
   });