+++ packages/sdk/react-client/src/hooks/bot-factory/useBotFactoryClient.ts
@@ -5,19 +5,21 @@
 import { useEffect, useState } from 'react';
 
 import { BotFactoryClient } from '@dxos/bot-factory-client';
+import { Config } from '@dxos/config';
 import { PublicKey } from '@dxos/crypto';
 import { raise } from '@dxos/debug';
+import { NetworkManager } from '@dxos/network-manager';
 
-import { useClient, useConfig } from '../client';
-
-export const useBotFactoryClient = (): BotFactoryClient | undefined => {
+export const useBotFactoryClient = (config: Config): BotFactoryClient | undefined => {
   const [botFactoryClient, setBotFactoryClient] = useState<BotFactoryClient>();
 
-  const client = useClient();
-  const config = useConfig();
-
   useEffect(() => {
-    const botFactoryClient = new BotFactoryClient(client.echo.networkManager);
+    const networkManager = new NetworkManager({
+      signal: config.get('runtime.services.signal.server') ? [config.get('runtime.services.signal.server')!] : undefined,
+      ice: config.get('runtime.services.ice'),
+      log: true
+    });
+    const botFactoryClient = new BotFactoryClient(networkManager);
     const topic = config.get('runtime.services.bot.topic') ??
       raise(new Error('Bot factory topic is not provided'));
     setImmediate(async () => {